Added helper classes for Player and Briefcase and extracted several methods from Program.cs
